+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This changelog will be used to generate documentation on [release notes page](ht
 - UseSampling and UseAdaptiveSampling extensions were moved to Microsoft.ApplicationInsights.Extensibility
 - Cut Phone and Store support
 - Updated ```DependencyTelemetry``` to have new properties ```ResultCode``` and ```Id```
-- If ``ServerTelemetryChannel`` is initialized programmatically it is required to call ServerTelemetryChannel.Initialize() method. Otherwise System.InvalidOperationException is thrown.
+- If ``ServerTelemetryChannel`` is initialized programmatically it is required to call ServerTelemetryChannel.Initialize() method. Otherwise persistent storage will not be initialized (that means that if telemetry cannot be sent because of temporary connectivity issues it will be dropped).
 - ``ServerTelemetryChannel`` has new property ``StorageFolder`` that can be set either through code or though configuration. If this property is set ApplicationInsights uses provided location to store telemetry that was not sent because of temporary connectivity issues. If property is not set or provided folder is inaccessible ApplicationInsights will try to use LocalAppData or Temp as it was done before.
 
 ## Version 2.0.0-beta3